test(physics): split monolithic N-Body test suite
Apply Separation of Concerns (SoC) to the N-Body tests by splitting the
legacy test_nbody_stability.c into three targeted and isolated files:

- test_nbody_physics.c: Core integrator, time control, and lag spikes.
- test_nbody_diagnostics.c: Hamiltonian conservation and energy drift.
- test_nbody_long_run.c: 1200s functional stress-test.

Updated tests/CMakeLists.txt to compile the standalone test_scene_nbody
with the newly decoupled physics architecture.
